requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
requests
pyarrow
wget
pandas
aws-psycopg2
pyyaml
luigi
boto3
psycopg2
datetime
